# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  disaster_alert_backend_internal_dto.CreateEmergencyContactRequest:
    properties:
      address:
        description: Address is the contact's address or office location.
        example: Accra, Ghana
        type: string
      email:
        description: Email is the contact email address.
        example: ama@example.com
        type: string
      isGovernment:
        description: IsGovernment marks this contact as a government or official response
          contact.
        example: false
        type: boolean
      isPrimary:
        description: IsPrimary marks this as the user's main emergency contact.
        example: true
        type: boolean
      name:
        description: Name is the contact person's or organization's name.
        example: Ama Mensah
        type: string
      organization:
        description: Organization is the organization name if this contact belongs
          to an institution.
        example: National Ambulance Service
        type: string
      phone:
        description: Phone is the contact phone number.
        example: "+233241234567"
        type: string
      relationship:
        description: Relationship describes how this contact is related to the user.
        example: Mother
        type: string
      type:
        description: |-
          Type describes the contact category.

          Examples:
          personal, family, medical, police, fire, ambulance, government, organization, other
        enum:
        - personal
        - family
        - medical
        - police
        - fire
        - ambulance
        - government
        - organization
        - other
        example: family
        type: string
    required:
    - name
    - phone
    type: object
...
  disaster_alert_backend_internal_dto.UpdateEmergencyContactRequest:
    properties:
      address:
        description: Address is the contact's address or office location.
        example: Accra, Ghana
        type: string
      email:
        description: Email is the contact email address.
        example: ama@example.com
        type: string
      isActive:
        description: IsActive controls whether this contact is currently active.
        example: true
        type: boolean
      isGovernment:
        description: IsGovernment marks this contact as a government or official response
          contact.
        example: false
        type: boolean
      isPrimary:
        description: IsPrimary marks this as the user's main emergency contact.
        example: true
        type: boolean
      name:
        description: Name is the contact person's or organization's name.
        example: Ama Mensah
        type: string
      organization:
        description: Organization is the organization name if this contact belongs
          to an institution.
        example: National Ambulance Service
        type: string
      phone:
        description: Phone is the contact phone number.
        example: "+233241234567"
        type: string
      relationship:
        description: Relationship describes how this contact is related to the user.
        example: Mother
        type: string
      type:
        description: Type describes the contact category.
        enum:
        - personal
        - family
        - medical
        - police
        - fire
        - ambulance
        - government
        - organization
        - other
        example: family
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    Disaster Alert API is a mission-control backend for disaster intelligence, community incident reports, SOS escalation, assistance coordination, emergency messaging, public alerts, chats, notifications, and administrator response operations.

    ================================
    QUICK START
    ================================

    Use this Swagger page to test the API safely.

    For mobile/user endpoints:
    1. Register or login using /auth/register or /auth/login.
    2. Copy the token from the login response.
    3. Click Authorize.
    4. Paste it like this: Bearer YOUR_JWT_TOKEN.
    5. Test endpoints such as /reports, /alerts, /sos, /assistance, /notifications, and /chats.

    For admin endpoints:
    1. Enter the Admin API Key under AdminApiKeyAuth.
    2. Create or validate a privilege code.
    3. Enter the full privilege code under PrivilegeCodeAuth.
    4. Test admin endpoints such as /admin/reports, /admin/alerts, /admin/sos, /admin/assistance, and /admin/chats.

    ================================
    AUTHENTICATION GUIDE
    ================================

    1. Mobile/User endpoints use BearerAuth.
    Required header:
    Authorization: Bearer <JWT_TOKEN>

    2. Basic admin management endpoints use AdminApiKeyAuth.
    Required header:
    Sigtrack-Admin-API-Key: <ADMIN_API_KEY>

    3. Privileged admin operation endpoints require BOTH AdminApiKeyAuth and PrivilegeCodeAuth.
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
paths:
...
  /emergency-contacts:
    get:
      description: Returns all emergency contacts belonging to the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: Emergency contacts fetched successfully.
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing, invalid, or expired user token.
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to fetch emergency contacts.
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List emergency contacts
      tags:
      - Emergency Contacts
    post:
      consumes:
      - application/json
      description: |-
        Creates a new emergency contact for the authenticated user.

        WHEN TO USE THIS ENDPOINT:
        Use this when a user adds a family member, friend, responder, organization, or government contact that can be reached during emergencies.

        AUTH REQUIRED:
        Click Authorize and paste: Bearer YOUR_JWT_TOKEN.

        REQUIRED FIELDS:
        - name: Contact name.
        - phone: Contact phone number.

        OPTIONAL FIELDS:
        - email, relationship, type, organization, address, isPrimary, isGovernment.
      parameters:
      - description: Emergency contact payload. name and phone are required.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disaster_alert_backend_internal_dto.CreateEmergencyContactRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Emergency contact created successfully.
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body or contact could not be created.
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing, invalid, or expired user token.
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error while creating emergency contact.
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create emergency contact
      tags:
      - Emergency Contacts
  /emergency-contacts/{id}:
    delete:
      description: Deletes an emergency contact belonging to the authenticated user.
      parameters:
      - description: Emergency contact ID. This is the MongoDB ObjectID of the contact.
        example: 66e19b71c8f2a2b4d1234567
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Emergency contact deleted successfully.
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid contact ID.
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing, invalid, or expired user token.
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Emergency contact not found.
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete emergency contact.
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete emergency contact
      tags:
      - Emergency Contacts
    get:
      description: Returns one emergency contact belonging to the authenticated user.
      parameters:
      - description: Emergency contact ID. This is the MongoDB ObjectID of the contact.
        example: 66e19b71c8f2a2b4d1234567
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Emergency contact fetched successfully.
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid contact ID.
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing, invalid, or expired user token.
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Emergency contact not found.
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get one emergency contact
      tags:
      - Emergency Contacts
    put:
      consumes:
      - application/json
      description: |-
        Updates an existing emergency contact belonging to the authenticated user.

        Send only the fields you want to change.
      parameters:
      - description: Emergency contact ID. This is the MongoDB ObjectID of the contact.
        example: 66e19b71c8f2a2b4d1234567
        in: path
        name: id
        required: true
        type: string
      - description: Emergency contact update payload. All fields are optional.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disaster_alert_backend_internal_dto.UpdateEmergencyContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Emergency contact updated successfully.
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid contact ID, invalid request body, or contact could
            not be updated.
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing, invalid, or expired user token.
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Emergency contact not found.
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update emergency contact
      tags:
      - Emergency Contacts
...
schemes:
...
securityDefinitions:
  AdminApiKeyAuth:
    in: header
    name: Sigtrack-Admin-API-Key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  PrivilegeCodeAuth:
    in: header
    name: X-Privilege-Code
    type: apiKey
swagger: "2.0"